requirement.txt
absl-py==1.4.0
attrs==23.1.0
base==0.0.0
blinker==1.6.2
celluloid==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
databases==0.7.0
dm-tree==0.1.8
docx==0.2.4
Flask==2.3.2
fonttools==4.40.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
jax==0.4.13
jaxlib==0.4.13
jaxopt==0.7
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.14.0
kiwisolver==1.4.4
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
neos==0.3.0
numpy==1.25.0
opt-einsum==3.3.0
optax==0.1.5
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pyhf==0.6.3
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
relaxed==0.2.2
requests==2.31.0
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
SQLAlchemy==1.4.48
toolz==0.12.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
